+++ composable/modal.ts
@@ -1,61 +1,67 @@
-import { ref } from "vue";
-import { isFunction } from "lodash";
+import { ref } from 'vue'
+import { isFunction } from 'lodash'
 
 export const useModal = () => {
-  const isOpen = ref<boolean>(false);
+  const isOpen = ref<boolean>(false)
 
   const open = (...callbacks: Function[]) => {
     callbacks.forEach((callback) => {
       if (isFunction(callback)) {
-        callback();
+        callback()
       }
-    });
-    isOpen.value = true;
-  };
+    })
+    isOpen.value = true
+  }
 
   const close = () => {
-    isOpen.value = false;
-  };
+    isOpen.value = false
+  }
 
   const toggle = () => {
-    isOpen.value = !isOpen.value;
-  };
+    isOpen.value = !isOpen.value
+  }
 
-  const invokeToggle = (...callbacks: Function[]) => {
+  const invokeToggle = (...callbacks: (Function | Promise<void>)[]) => {
     callbacks.forEach((callback) => {
       if (isFunction(callback)) {
-        callback();
+        callback()
       }
-    });
-    isOpen.value = !isOpen.value;
-  };
+    })
+    isOpen.value = !isOpen.value
+  }
 
-  const invokeAsyncToggle = async (...callbacks: Function[]) => {
+  const invokeAsyncToggle = async (...callbacks: (Function | Promise<void>)[]) => {
     for (const callback of callbacks) {
       if (isFunction(callback)) {
-        await callback();
+        await callback()
+      } else if (callback instanceof Promise) {
+        await callback
       }
     }
-    isOpen.value = !isOpen.value;
-  };
+    isOpen.value = !isOpen.value
+  }
 
-  const toggleInvoke = (...callbacks: Function[]) => {
-    isOpen.value = !isOpen.value;
+  const toggleInvoke = (...callbacks: (Function | Promise<void>)[]) => {
+    isOpen.value = !isOpen.value
     callbacks.forEach((callback) => {
       if (isFunction(callback)) {
-        callback();
+        callback()
+      } else if (callback instanceof Promise) {
+        callback
       }
-    });
-  };
+    })
+  }
 
-  const toggleInvokeAsync = async (...callbacks: Function[]) => {
-    isOpen.value = !isOpen.value;
+  const toggleInvokeAsync = async (...callbacks: (Function | Promise<void>)[]) => {
+    isOpen.value = !isOpen.value
     for (const callback of callbacks) {
       if (isFunction(callback)) {
-        await callback();
+        await callback()
+      } else if (callback instanceof Promise) {
+        await callback
       }
     }
-  };
+  }
 
   return {
     isOpen,
@@ -65,6 +71,6 @@ export const useModal = () => {
     invokeToggle,
     toggleInvoke,
     invokeAsyncToggle,
-    toggleInvokeAsync,
-  };
-};
+    toggleInvokeAsync
+  }
+}