+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [0.0.9] - 2024-06-16
+### Fixed
+- fix RouterView and Transition
+
 ## [0.0.8] - 2024-06-01
 ### Fixed
 - table fetch method to post